% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ.cv.R
\name{circ.cv}
\alias{circ.cv}
\title{Circular Curves}
\usage{
circ.cv(n, cen = matrix(c(0, 0), byrow = TRUE, ncol = 2), r = 1)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}

\item{cen}{a \eqn{m*2} matrix where each of its row specifies a centre of the circle (by default only the origin (0,0)).}

\item{r}{(by default \eqn{1}) the radius of the circles.}
}
\description{
Creates a bivariate sample belonging to the circular curves dependence structure
}
\examples{
X <- circ.cv(1000,matrix(c(1,1,-1,-1,0,0), byrow=TRUE, ncol=2))
qplt(X)
}
\references{
Geenens G. & Lafaye de Micheaux P., (2018),
The Hellinger Correlation,
arXiv:1810.10276.
}
