% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computepowers.R
\name{computepowers}
\alias{computepowers}
\title{Compute Powers of All Competitors}
\arguments{
\item{features}{features obtained from importfeatures().}

\item{model}{string indicating the model ('allnet' or 'allCNN')}

\item{Model}{keras model object}

\item{n}{sample size.}
}
\description{
Compute Powers of All Competitors
}
