% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre.cld.R
\name{cre.cld}
\alias{cre.cld}
\title{Crescent Clouds}
\usage{
cre.cld(n, c = 0.5)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}

\item{c}{(by default \eqn{0.5}) waxes/wanes of the crescent shape.}
}
\value{
matrix X with two columns representing the bivariate sample.
}
\description{
Creates a bivariate sample belonging to the crescent clouds dependence structure
}
\examples{
X <- cre.cld(1000)
qplt(X)
}
