% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.cv.R
\name{cross.cv}
\alias{cross.cv}
\title{Cross Curves}
\usage{
cross.cv(n, g = 1)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}

\item{g}{(by default \eqn{1}) angle of rotation to obtain the cross shape.}
}
\description{
Creates a bivariate sample belonging to the cross curves dependence structure
}
\examples{
X <- cross.cv(1000)
qplt(X)
}
\references{
Geenens G. & Lafaye de Micheaux P., (2018),
The Hellinger Correlation,
arXiv:1810.10276.
}
