% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic.cv.R
\name{cubic.cv}
\alias{cubic.cv}
\title{Cubic Curves}
\usage{
cubic.cv(n)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}
}
\description{
Creates a bivariate sample belonging to the cubic curves dependence structure
}
\examples{
X <- cubic.cv(1000)
qplt(X)
}
\references{
Geenens G. & Lafaye de Micheaux P., (2018),
The Hellinger Correlation,
arXiv:1810.10276.
}
