% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexpowers.R, R/plotpowers.R
\docType{data}
\name{defaultcol}
\alias{defaultcol}
\title{LaTex Table Codes Printing}
\format{
An object of class \code{character} of length 28.

An object of class \code{character} of length 28.
}
\usage{
defaultcol

defaultcol
}
\arguments{
\item{PowersSummary}{powers summary table from powerssummary().}

\item{rows}{row names.}

\item{cuts}{indices for table separation.}

\item{decimals}{number of decimals to be printed.}

\item{Powers}{powers data matrix either from computepowers() or attached data in the package.}

\item{title}{title of the plot, defaults to 'Power Curves'.}

\item{special}{competitors with highlighted power curves.}

\item{cols}{column names.}

\item{nsample}{a vector of sample sizes, defaults to c(30,50,100,200,300,400).}
}
\description{
LaTex Table Codes Printing

Power Plots
}
\examples{
plotpowers(addPowers[11+seq(0,65,11),],'Additional Test 7')
plotpowers(addPowers[13+seq(0,79,14),],'Additional Test 7 A')
}
\keyword{datasets}
