% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depgen.R
\name{depgen}
\alias{depgen}
\title{Dependent Sample Generation}
\usage{
depgen(num, n, expr, randrotate = TRUE, print = FALSE)
}
\arguments{
\item{num}{number of repetitions applied.}

\item{n}{size of the sample (number of bivaraite data pairs).}

\item{expr}{a string describing the code to generate a dependent sample.}

\item{randrotate}{(TRUE/FALSE) default being TRUE to rotate the sample randomly.}

\item{print}{(TRUE/FALSE) default being FALSE to print out the number progress.}
}
\description{
Dependent Sample Generation
}
