% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diam.cld.R
\name{diam.cld}
\alias{diam.cld}
\title{Diamond Clouds}
\usage{
diam.cld(n, theta = pi/4)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}

\item{theta}{(by default \eqn{\pi}) angle of rotation to obtain the diamond shape.}
}
\value{
matrix X with two columns representing the bivariate sample.
}
\description{
Creates a bivariate sample belonging to the diamond clouds dependence structure
}
\examples{
X <- diam.cld(1000)
qplt(X)
}
\references{
Geenens G. & Lafaye de Micheaux P., (2018),
The Hellinger Correlation,
arXiv:1810.10276.
}
