% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doppler.cv.R
\name{doppler.cv}
\alias{doppler.cv}
\title{Doppler Curves}
\usage{
doppler.cv(n, f = 4)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}
}
\description{
Creates a bivariate sample belonging to the doppler curves dependence structure
}
\examples{
X <- doppler.cv(1000)
qplt(X)
}
\references{
Geenens G. & Lafaye de Micheaux P., (2018),
The Hellinger Correlation,
arXiv:1810.10276.
}
