% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.cv.R
\name{exp.cv}
\alias{exp.cv}
\title{Exponential Curves}
\usage{
\method{exp}{cv}(n, b = 2)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}

\item{b}{(by default \eqn{2}) specifies the exponential base.}
}
\description{
Creates a bivariate sample belonging to the exponential curves dependence structure
}
\examples{
X <- exp.cv(1000)
qplt(X)

}
