% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurereshape.R
\name{featurereshape}
\alias{featurereshape}
\title{Feature Reshaping (for keras models)}
\arguments{
\item{features}{features obtained from importfeatures().}

\item{model}{string indicating the model ('allnet' or 'allCNN')}
}
\description{
Feature Reshaping (for keras models)
}
