% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavysine.cv.R
\name{heavysine.cv}
\alias{heavysine.cv}
\title{Heavy Sine Curves}
\usage{
heavysine.cv(n, a = -0.6, b = -0.2, c = 0.2, d = 0.6)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}

\item{a, b, c, d}{specifies the cut of heavy sines (by default \eqn{-0.6,-0.2,0.2,0.6}).}
}
\description{
Creates a bivariate sample belonging to the heavy-sine curves dependence structure
}
\examples{
X <- heavysine.cv(1000)
qplt(X)
}
\references{
Jiang H. J. & Wu Q. L., (2018),
Robust Dependence Measure for Detecting Associations in Large Data Set,
Acta. Math. Sci., 38(1), 931-949.
}
