% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importfeatures.R
\name{importfeatures}
\alias{importfeatures}
\title{Feature Importation}
\usage{
importfeatures(scores, image, n)
}
\arguments{
\item{scores}{txt file with competitor scores.}

\item{image}{txt file with image greyscale pixel values.}

\item{n}{sample size.}
}
\description{
Feature Importation
}
