% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indgen.R
\name{indgen}
\alias{indgen}
\title{Independent Uniform Bivariate Sample Generator}
\usage{
indgen(n, sizes = c(50, 100, 200, 500, 1000))
}
\arguments{
\item{n}{numbers of replication produced.}

\item{size}{a vector of sample sizes}
}
\description{
Independent Uniform Bivariate Sample Generator
}
\examples{
X <- indgen(n=10, sizes = c(50,100,200,500,1000))
#10 copies of sample sizes 50,100,200,500,1000 each created
#statistically the same as
X <- indgen(n=10*sum(50,100,200,500,1000), sizes = c(1))
}
