% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indsim.R
\name{indsim}
\alias{indsim}
\title{Random Estimates of Measures over Independent Samples}
\usage{
indsim(f, n = 100, runs = 1000)
}
\arguments{
\item{f}{function of any measure or dependence test in the package.}

\item{n}{size of the sample (number of bivaraite data pairs).}

\item{runs}{numbers of repetitions applied.}
}
\description{
Random Estimates of Measures over Independent Samples
}
