% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexpowers.R
\docType{data}
\name{latexpowers}
\alias{latexpowers}
\alias{defaultcol}
\title{Print Power Tables for LaTex}
\format{
An object of class \code{character} of length 29.
}
\usage{
defaultcol
}
\arguments{
\item{Powers}{the matrix of powers of competitors (cols) over test scenarios (rows).}

\item{Group}{names of the test scenarios.}

\item{sizes}{a vector of smaple sizes.}

\item{nspecial}{number of models assessed.}

\item{decimals}{number of decimals to be printed.}
}
\description{
This function is prints power tables for LaTex.
}
\keyword{datasets}
