% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localdep.R
\name{localdep}
\alias{localdep}
\title{Local Dependence Generator}
\usage{
localdep(X, prop)
}
\arguments{
\item{X}{the matrix of random sample.}

\item{prop}{proportion of the sample randomised.}
}
\description{
Takes a sample and add noises to the sides
}
\examples{
X <- cross.cv(1000)
qplt(localdep(X,0.75))

}
