% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mace.R
\name{mace}
\alias{mace}
\title{Alternating Conditional Expectations}
\usage{
mace(X)
}
\description{
Alternating Conditional Expectations
}
\references{
Spector P.; Friedman J.; Tibshirani R.; Lumley T.; Garbett S.; Baron J., (2019), ACE and AVAS for Selecting Multiple Regression Transformations, R package, Version 1.4.1.
}
