% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mauk.R
\name{mauk}
\alias{mauk}
\title{Area Under Kendall Plot}
\usage{
mauk(X)
}
\description{
Area Under Kendall Plot
}
\references{
Vexler A.; Chen X.; Hutson A. D., (2017), Dependence and Independence Structure and Inference, StatMethodsMed Res., 2017 Oct;26(5):2114-2132.
}
