% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdist.R
\name{mdist}
\alias{mdist}
\title{Distance Correlation}
\usage{
mdist(X)
}
\description{
Distance Correlation
}
\references{
Richards D. S. P., (2017), Distance Correlation: A New Tool for Detecting Association and Measuring Correlation between Data Sets, arXiv:1709.06400.
}
