% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medc.R
\name{medc}
\alias{medc}
\title{Estimated Copula Dependence Measure}
\usage{
medc(X)
}
\description{
Estimated Copula Dependence Measure
}
\references{
Jiang H. J. &Wu Q. L., (2018), Robust Dependence Measure for Detecting Associations in Large Data Set, Acta.Math. Sci., 38(1), 931-949.
}
