% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhell.R
\name{mhell}
\alias{mhell}
\title{Hellinger Correlation}
\usage{
mhell(X)
}
\description{
Hellinger Correlation
}
\references{
Geenens G. & Lafaye de Micheaux P., (2018), The Hellinger Correlation, arXiv:1810.10276.
}
