% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhoeff.R
\name{mhoeff}
\alias{mhoeff}
\title{Hoeffding Dependence Coefficient}
\usage{
mhoeff(X)
}
\description{
Hoeffding Dependence Coefficient
}
\references{
Hoeffding W., (1948), A Non-Parametric Test of Independence, Ann. Math. Statist., 19(4), 1948, 546-557.
}
