% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhsic.R
\name{mhsic}
\alias{mhsic}
\title{Hilbert Schmidt Independence Criterion}
\usage{
mhsic(X, k = "gaussian")
}
\description{
Hilbert Schmidt Independence Criterion
}
\references{
Gretton A.; Bousquet O.; Smola A.; Schölkopf B., (2005), Measuring Statistical Dependence with Hilbert-Schmidt Norms, ALT’05 Proceedings of the 16th international conference on Algorithmic Learning Theory, 63-77.
}
