% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minf.R
\name{minf}
\alias{minf}
\title{Information Correlation Coefficient}
\usage{
minf(X, b1 = 10, b2 = 10)
}
\description{
Information Correlation Coefficient
}
\references{
Hausser J. & Strimmer K., (2015), Estimation of Entropy, Mutual Information and Related Quantities, R package, Version 1.2.1.
}
