% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmdif.R
\name{mmdif}
\alias{mmdif}
\title{Martingale Difference Correlation}
\usage{
mmdif(X, c = "U")
}
\description{
Martingale Difference Correlation
}
\references{
Shao X. & Zhang J., (2014), Martingale Difference Correlation and Its Use in High-Dimensional Variable Screening, Journal of the American Statistical Association, 109(507), 2014.
}
