% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmic.R
\name{mmic}
\alias{mmic}
\title{Maximal Information Coefficient}
\usage{
mmic(X)
}
\description{
Maximal Information Coefficient
}
\references{
Filosi M.; Visintainer R.; Albanese D.; Riccadonna S.; Jurman G.; Furlanello C., (2019), Maximal Information-Based Nonparametric Exploration for Variable Analysis, R package, Version 1.5.8.
}
