% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{model_predict}
\alias{model_predict}
\title{Model Prediction}
\usage{
model_predict(model, model_name, df)
}
\arguments{
\item{model}{the model object.}

\item{model_name}{a string of model name, must be either `'image'`, `'score'` or `'combined'.`}

\item{df}{the input data for model prediction.}
}
\description{
This function is used to compute model predictions from either image, score or combined model.
}
