% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrand.R
\name{mrand}
\alias{mrand}
\title{Randomised Dependence Coefficient}
\usage{
mrand(X, k = 10, s = 1/2, Phi = c(sin, cos), seed)
}
\description{
Randomised Dependence Coefficient
}
\references{
Lopez-PazD.;Hennig P.; Schölkopf B., (2013) The Randomised Dependence Coefficient, arXiv:1304.7717.
}
