% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normnoise.R
\name{normnoise}
\alias{normnoise}
\title{Adding Normally Distributed Noises}
\usage{
normnoise(X, sigma, both = FALSE)
}
\arguments{
\item{X}{the matrix of random sample.}

\item{sigma}{the standard deviation of the normal noise added.}
}
\description{
Adding Normally Distributed Noises
}
