% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parabola.cv.R
\name{parabola.cv}
\alias{parabola.cv}
\title{Parabola Curves}
\usage{
parabola.cv(n)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}
}
\description{
Creates a bivariate sample belonging to the parabola curves dependence structure
}
\examples{
X <- parabola.cv(1000)
qplt(X)
}
