% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pddr.R
\name{pddr}
\alias{pddr}
\title{Data--Driven Rank Test}
\usage{
pddr(X, num.MC = 100)
}
\description{
Data--Driven Rank Test
}
\references{
Kallenberg W. C. M.; Ledwina T., (1999) Data-Driven Rank Tests for Independence, Journal of the American Statistical Association, 94(445), 285-301.
}
