% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phhg.R
\name{phhg}
\alias{phhg}
\title{Heller-Heller-Gorfine Test}
\usage{
phhg(X, nperm = 50)
}
\description{
Heller-Heller-Gorfine Test
}
\references{
Heller R.; Heller Y; GorfineM., (2012), A Consistent Multivariate Test of Association Based on Ranks of Distances, arXiv:1201.3522.
}
