% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpowers.R
\name{plotpowers}
\alias{plotpowers}
\title{Plot Power Curves}
\usage{
plotpowers(power_data, title)
}
\arguments{
\item{power_data}{data table of the powers.}

\item{title}{a vector of titles.}
}
\description{
This function is used to produce power plots
}
