% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randimage.R
\name{randimage}
\alias{randimage}
\title{Sample Generator from Image}
\usage{
randimage(n, image)
}
\arguments{
\item{n}{size of the sample (number of bivaraite data pairs).}

\item{image}{an image object (under imager package).}
}
\description{
Produce a sample with probability determined by greyscale pixels of an image
}
