% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randlin.R
\name{randlin}
\alias{randlin}
\title{Linear Lines with Random Gradient}
\usage{
randlin(n, a = -2, b = 2)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}

\item{a}{lower bound of the gradient.}

\item{b}{upper bound of the gradient.}
}
\description{
Linear Lines with Random Gradient
}
