% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{The Rotation Function}
\usage{
rotate(X, theta)
}
\arguments{
\item{X}{the matrix of random sample.}

\item{theta}{radian of rotation.}
}
\description{
Anti-clockwise rotation of a sample around the origin
}
