% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleapply.R
\name{sampleapply}
\alias{sampleapply}
\title{Applications over Samples}
\arguments{
\item{X}{the matrix of random sample.}

\item{size}{a vector of sample sizes}

\item{option}{(0/1) 0 for greyscale computation and 1 for competitor scores computation.}

\item{grid}{dimension of the square grid of grescale image pixels.}

\item{printall}{(TRUE/FALSE) default being FALSE would give a progress bar, if TRUE prints out the number progress.}
}
\description{
This function is used to compute features (competitor scores or image greyscale pixels)
}
