% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtaegeuk.cv.R
\name{samtaegeuk.cv}
\alias{samtaegeuk.cv}
\title{Samtaegeuk Curves}
\usage{
samtaegeuk.cv(n)
}
\arguments{
\item{n}{size of the sample (number of bivaraite data pairs).}
}
\description{
Creates a bivariate sample belonging to the samtaegeuk curves dependence structure
}
\examples{
X <- samtaegeuk.cv(1000)
qplt(X)
}
