% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sine.cv.R
\name{sine.cv}
\alias{sine.cv}
\title{Sine Curves}
\usage{
sine.cv(n, s = 2)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}

\item{s}{frequency of sine waves.}
}
\description{
Creates a bivariate sample belonging to the sine curves dependence structure
}
\examples{
X <- sine.cv(1000)
qplt(X)
}
\references{
Geenens G. & Lafaye de Micheaux P., (2018),
The Hellinger Correlation,
arXiv:1810.10276.
}
