% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral.cv.R
\name{spiral.cv}
\alias{spiral.cv}
\title{Spiral Curves}
\usage{
spiral.cv(n, k = 2)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}

\item{k}{intensity of spirals.}
}
\description{
Creates a bivariate sample belonging to the spiral curves dependence structure
}
\examples{
X <- spiral.cv(1000)
qplt(X)
}
\references{
Jiang H. J. & Wu Q. L., (2018),
Robust Dependence Measure for Detecting Associations in Large Data Set,
Acta. Math. Sci., 38(1), 931-949.
}
