% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taegeuk.cv.R
\name{taegeuk.cv}
\alias{taegeuk.cv}
\title{Taegeuk Curves}
\usage{
taegeuk.cv(n)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}
}
\description{
Creates a bivariate sample belonging to the taegeuk curves dependence structure
}
\examples{
X <- taegeuk.cv(1000)
qplt(X)
}
\references{
Jiang H. J. & Wu Q. L., (2018),
Robust Dependence Measure for Detecting Associations in Large Data Set,
Acta. Math. Sci., 38(1), 931-949.
}
