% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri.cld.R
\name{tri.cld}
\alias{tri.cld}
\title{Triangle Clouds}
\usage{
tri.cld(n, l = 1, r = 1)
}
\arguments{
\item{n}{size of the sample (number of bivaraite data pairs).}

\item{l, r}{extent of span towards the left and the right of the triangle.}
}
\description{
Creates a bivariate sample belonging to the triangle clouds dependence structure
}
\examples{
X <- tri.cld(1000)
qplt(X)
}
