% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoparabola.cv.R
\name{twoparabola.cv}
\alias{twoparabola.cv}
\title{Two Parabolae Curves}
\usage{
twoparabola.cv(n)
}
\arguments{
\item{n}{size of the sample (number of bivaraite data pairs).}
}
\description{
Creates a bivariate sample belonging to the two-parabolae dependence structure
}
\examples{
X <- twoparabola.cv(1000)
qplt(X)
}
\references{
Geenens G. & Lafaye de Micheaux P., (2018),
The Hellinger Correlation,
arXiv:1810.10276.
}
