% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varnormnoise.R
\name{varnormnoise}
\alias{varnormnoise}
\title{Adding Normally Ditributed Noises with Non-Constant Variance}
\usage{
varnormnoise(X, expr = "0.1")
}
\arguments{
\item{X}{the matrix of random sample.}

\item{expr}{a string with its absolute value describing the variance function.}
}
\description{
Adding Normally Ditributed Noises with Non-Constant Variance
}
