% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.cv.R
\name{w.cv}
\alias{w.cv}
\title{W-Shape Curves}
\usage{
w.cv(n)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}
}
\description{
Creates a bivariate sample belonging to the w-shape dependence structure
}
\examples{
X <- w.cv(1000)
qplt(X)
}
\references{
Geenens G. & Lafaye de Micheaux P., (2018),
The Hellinger Correlation,
arXiv:1810.10276.
}
