% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedge.cv.R
\name{wedge.cv}
\alias{wedge.cv}
\title{Wedge Curves}
\usage{
wedge.cv(n)
}
\arguments{
\item{n}{size of the sample (number of bivariate data pairs).}
}
\description{
Creates a bivariate sample belonging to the wedge dependence structure
}
\examples{
X <- wedge.cv(1000)
qplt(X)
}
\references{
Geenens G. & Lafaye de Micheaux P., (2018),
The Hellinger Correlation,
arXiv:1810.10276.
}
